Automation scripts

+++ hsSphere_rcf/Allclean.sh
@@ -1,8 +1,32 @@
 #!/bin/bash
 cd "$(dirname "$0")" # Change directory to host directory for runtime of script
 
+while getopts zp OPTION # colon (:) requires an input with -p
+do
+    case ${OPTION} in
+    z)
+        KEEPZERO='true'
+        print "Running with mesher enabled"
+        ;;
+    p)
+        KEEP_PM='true'
+        print "Running with mesher enabled"
+        ;;
+    ?)
+        echo "Invalid input argument" >&2
+        usage
+        exit 1
+        ;;
+    esac
+done
+
 # List of directories to keep
+
 KEEP_DIRS=("0.orig" "constant" "system") 
+if [[ "${KEEPZERO}" = 'true' ]]
+then
+    KEEP_DIRS=("0.orig" "constant" "system" "0") 
+fi
 
 # Loop through all directories
 for dir in */; do
@@ -17,4 +41,7 @@ for dir in */; do
 done
 
 # delete constant/polyMesh
-rm -rf "constant/polyMesh"
+if [[ "${KEEP_PM}" != 'true' ]]
+then
+    rm -rf "constant/polyMesh"
+fi
